refactor: dev,prod 쿠키 분리
