+++ apps/web/src/app/mypage/_components/Footer/_clientBoundary/SignoutClient/SignoutClient.tsx
@@ -5,7 +5,8 @@ import usePopup from '@/hooks/usePopup';
 import Modal from '@/components/popup/modal/Modal';
 import TextButton from '@/components/buttons/textButton/TextButton';
 import { ROUTES } from '@/constants/routes';
-import { COUPLE_STATUS, COUPLE_STATUS_COOKIE } from '@/constants/coupleStatus';
+import { COUPLE_STATUS_COOKIE } from '@/constants/cookie';
+import { COUPLE_STATUS } from '@/constants/coupleStatus';
 import * as S from './SignoutClient.styles';
 
 function getCookieValue(name: string): string | undefined {
