+++ apps/web/src/middleware.ts
@@ -1,11 +1,12 @@
 import { NextRequest, NextResponse } from 'next/server';
 
 import { ROUTES } from '@/constants/routes';
-import { COUPLE_STATUS, COUPLE_STATUS_COOKIE } from '@/constants/coupleStatus';
+import { ACCESS_TOKEN_COOKIE, COUPLE_STATUS_COOKIE } from '@/constants/cookie';
+import { COUPLE_STATUS } from '@/constants/coupleStatus';
 
 export function middleware(request: NextRequest) {
   const { pathname } = request.nextUrl;
-  const accessToken = request.cookies.get('accessToken')?.value;
+  const accessToken = request.cookies.get(ACCESS_TOKEN_COOKIE)?.value;
   const coupleStatus = request.cookies.get(COUPLE_STATUS_COOKIE)?.value;
   const isLoginPage = pathname.startsWith(ROUTES.LOGIN);
